requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.0
supabase>=2.3.0
openai>=1.12.0
python-dotenv>=1.0.0
twscrape>=0.12.0
feedparser>=6.0.0
trafilatura>=1.6.0
lxml>=5.0.0